requirements.txt
git+https://github.com/xzzy/webtemplate.git#egg=webtemplate_dbca
git+https://github.com/dbca-wa/dbca-utils@1.1.1#egg=dbca-utils
psycopg2-binary==2.8.5
Django==3.2.12
django-extensions==1.9.8
ipython==7.30.1
social-auth-app-django==2.1.0
git+https://github.com/xzzy/ledger_api_client.git#egg=ledger_api_client
#git+https://github.com/xzzy/ledger_api_client.git@c6d6fef2c82e3fc821d4132ffbcfa81c98929047#egg=ledger_api_client
git+https://github.com/scottp-dpaw/social-core.git@email_fix#egg=social-auth-core
coverage==4.3.1
coveralls==1.1
reportlab==3.5.26
#django_bootstrap3==15.0.0
django-braces>=1.8.1
django-datatables-view==1.13.0
django-reversion==4.0.1
django-reversion-compare==0.14.1
git+https://github.com/xzzy/django-preserialize.git#egg=django-preserialize
django-countries==7.2.1
django-cron==0.5.0
django-dynamic-fixture==1.9.1
openpyxl==3.0.7
datapackage==1.15.2
python-dateutil==2.6.1
py4j==0.10.2.1
djangorestframework==3.12.0
#djangorestframework-gis==0.16
djangorestframework-gis==0.18
djangorestframework-csv==2.0.0
pycountry==17.1.8
six>=1.10.0
django-ical==1.4
django-taggit==1.3.0
geojson==1.3.3
unicodecsv==0.14.1
PyPDF2==1.26.0
django-extra-fields==0.9
django-crispy-forms==1.7.0
django-reset-migrations==0.3.1
django-ckeditor==5.4.0
djangorestframework-datatables==0.4.0
django-confy==1.0.4
xlwt==1.3.0
xlsxwriter==1.1.2
gunicorn==19.9.0
dj-static==0.0.6
Pillow==7.2.0
mixer==6.1.3
django-dirtyfields==1.3.1
python-docx==0.8.10
psutil==5.7.0
git+https://github.com/xzzy/django_site_queue.git#egg=django_site_queue
dj-database-url==0.5.0
phonenumbers==8.12.15
django-treebeard==4.3.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
django-phonenumber-field==5.2.0
gevent==21.1.2
whitenoise==5.3.0
django-multiselectfield==0.1.12
